+++ packages/calendar/src/components/week-grid/__test__/time-grid-day/utils.tsx
@@ -3,15 +3,21 @@ import { CalendarEventInternal } from '@schedule-x/shared/src/interfaces/calenda
 import { render } from '@testing-library/preact'
 import { AppContext } from '../../../../utils/stateful/app-context'
 import TimeGridDay from '../../time-grid-day'
+import { BackgroundEvent } from '@schedule-x/shared/src/interfaces/calendar/background-event'
 
 export const renderComponent = (
   $app: CalendarAppSingleton,
   calendarEvents: CalendarEventInternal[],
-  date: string
+  date: string,
+  backgroundEvents: BackgroundEvent[] = []
 ) => {
   render(
     <AppContext.Provider value={$app}>
-      <TimeGridDay calendarEvents={calendarEvents} date={date} />
+      <TimeGridDay
+        calendarEvents={calendarEvents}
+        date={date}
+        backgroundEvents={backgroundEvents}
+      />
     </AppContext.Provider>
   )
 }