+++ packages/translations/src/locales/en-GB/date-picker.ts
@@ -6,5 +6,4 @@ export const datePickerEnGB: DatePickerTranslations = {
   'Next month': 'Next month',
   'Previous month': 'Previous month',
   'Choose Date': 'Choose Date',
-  'Select View': 'Select View',
 }