+++ packages/translations/src/locales/fr-FR/date-picker.ts
@@ -6,5 +6,4 @@ export const datePickerFrFR: DatePickerTranslations = {
   'Next month': 'Mois suivant',
   'Previous month': 'Mois précédent',
   'Choose Date': 'Choisir une date',
-  'Select View': 'Choisir la vue',
 }